requirements.txt
python-telegram-bot
supabase
dotenv
pycryptodome